python/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
asyncio==3.4.3
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
motor==3.6.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
